refactor(hir): Split class_dict.rs
